Enabled scanning for devices

+++ process-viewer/Controller/BluetoothCtl.swift
@@ -12,7 +12,7 @@ class BluetoothCtl : ObservableObject {
     @Published var ssh: SSH
     @Published var isAvailable: Bool
     @Published var btDevices: [BTDevice]
-    var scanTaskHandle: Task<Void, Error>?
+    @Published var scanTaskHandle: Task<Void, Error>?
     
     init(ssh: SSH) {
         self.ssh = ssh
@@ -26,6 +26,10 @@ class BluetoothCtl : ObservableObject {
                 let stdout = try await self.ssh.runSync(cmd: "bluetoothctl show")
                 
                 self.isAvailable = !stdout.contains("No default controller available")
+                
+                if self.isAvailable {
+                    self.scanOn()
+                }
             } catch {
                 print("Could not run command: \(error)")
                 self.isAvailable = false
@@ -34,6 +38,10 @@ class BluetoothCtl : ObservableObject {
     }
     
     func scanOn() {
+        if !self.isAvailable {
+            return
+        }
+        
         do {
             self.scanTaskHandle = try self.ssh.runAsync(cmd: "bluetoothctl scan on") { stdout in
                 print("Stdout: \(stdout)")
@@ -46,6 +54,10 @@ class BluetoothCtl : ObservableObject {
     }
     
     func scanOff() {
+        if !self.isAvailable {
+            return
+        }
+        
         Task.detached { @MainActor in
             do {
                 let _ = try await self.ssh.runSync(cmd: "pkill -f 'bluetoothctl scan on'")
@@ -53,6 +65,7 @@ class BluetoothCtl : ObservableObject {
                 print("Error when trying to disable scanning: \(error)")
             }
             self.scanTaskHandle?.cancel()
+            self.scanTaskHandle = nil
         }
     }
 }
